use int64_t bobo

+++ src/core/vdp.hpp
@@ -7,6 +7,7 @@
 #ifndef INCLUDE_VDP_HPP
 #define INCLUDE_VDP_HPP
 #include <stdio.h>
+#include <stdint.h>
 #include <string.h>
 
 class VDP
@@ -64,7 +65,7 @@ class VDP
   public:
     unsigned short display[240 * 192];
     struct Context {
-        int bobo;
+        int64_t bobo;
         int countV;
         int countH;
         unsigned char status;
